Use Status instead of DisconnectReason in event
According to the spec, the reason field of the event may have any valid status code. For example, if host initiates a disconnect, the controller will generate a disconnect event with a reason outside the DisconnectReason enum. Changing it to use the Status type will ensure it can handle any value the controller is allowed to send.

The earlier approach was to extend the DisconnectReason with values returned by the controller, but this allows the user to break the spec wrt. which reasons can be
send in the Disconnect command. Therefore, the DisconnectReason enum now only allow the values as defined by the spec.