use correct bitmask for decoding acl handle
Fixes a bug with the bit mask used to select the bits relevant for the ACL handle. The old mask of 0xeff refers to the max value, and used as a bitmask will mask the 9th bit. Using the full 12 bit bitmask fixes this, and the value being <= 0xeff is checked in the ConnHandle::new() method.